msm: vidc: Adding check for unsuccessful CMA enablement and VMID switch
Catching the return value for the function of enabling CMA and
checking if the CMA enablement is successful and also the VMID switch.
On failure of VMID switch during CMA enablement, the VMID switch won't
be called again during normal usecase and the previous context banks
will be attached again.

Change-Id: I9f1cad1df14f0e6502efabf79fea2e9d92a31c3e
Signed-off-by: Sayantan Majumder <smajumde@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: firebird11 <hbgassel@gmail.com>